Refactor fully to ES6 imports + linting
